Use OpenCGA client from jsorolla
Client is written for vanilla JS, use Node VM to parse and execute file
in context. Use dummy RestClient to proxy requests rather than the
provided one.

For the sake of the prototype, login straight away and pull the
password from the environment variables.

Run node with Harmony flag to correctly parse OpenCGAClient which is
written in ES6.

https://nodejs.org/api/vm.html